# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAtext
  description:  |
    This API allow multiple users to chat with each other,
    one-to-one or by creating group chat, an user can start
    new conversation by searching other user, or add them to
    an existing chat to create a group.
  version: 0.2.1fray
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  parameters:
    usr_id:
      description:  |
        Unique user id
      
      schema:
        $ref: "#/components/schemas/User/properties/usrId"
      name: usr_id
      in: path
      required: true
    chat_id:
      description:  |
        Unique chat id
      
      schema:
        $ref: "#/components/schemas/Chat/properties/chatId"
      name: chat_id
      in: path
      required: true
...
  schemas:
    User:
      description:  |
        Contain information about an user of the app,
        such as its name, user id, and profile photo.
      
      type: object
      title: User

      properties:
        usrId:
          description:  |
            Unique user id
      
          type: string
          pattern: '^[^\s].*[^\s]$'
          example: "ofnarc754"
          
          minLength: 3
          maxLength: 32
          readOnly: true
        
        userName:
          description:  |
            Username
          
          type: string
          pattern: '^[^\s].*[^\s]$'
          example: "franco"
          
          minLength: 3
          maxLength: 16
        
        userPhoto:
          description:  |
            User profile photo (propic) encoded in base64
          
          type: string
          format: binary
          example: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
          
          minLength: 0
          maxLength: 1398101 #Equivalente di 1MB di dati binari in base64
    Chat:
      description:  |
        Contain information about a chat of two or more users,
        such as the name of the chat, the propic the participants
        and the id.

      type: object
      title: Chat
      
      properties:
        chatId:
          description:  |
            Unique chat id
          
          type: integer
          
          minimum: 0
          maximum: 999
          readOnly: true

        chatName:
          description:  |
            Chat name, if it's a one-to-one chat the name
            correspond to the other username
          
          type: string
          pattern: '^[^\s].*[^\s]$'
          example: "Chat Uni"
          
          minLength: 3
          maxLength: 16

        chatPhoto:
          description:  |
            Chat Photo, if it's a one-to-one chat is the propic of the other user,
            else contain the group photo
          
          type: string
          format: binary
          example: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."

          minLength: 0
          maxLength: 1398101

        isGroup:
          description:  |
            Used to distinguish group chat between one-to-one chat_id.
            false: one-to-one chat with only 2 users
            true: chat with more than 2 users
            
          type: boolean

        participants:
          description:  |
            Array containing the users id of the users in the chat
          
          type: array
          minItems: 2
          maxItems: 64
          items:
            $ref: "#/components/schemas/User/properties/usrId"
...
    Error:
      description: Personalizable error
      type: object
      properties:
        content:
          description: Short description of the error

          type: string
          pattern: "^.*$"

          minLength: 3
          maxLength: 40
          example: "token expired"

        code:
          description: Error code

          type: string
          pattern: "^[A-Z0-9 ]+$"

          minLength: 3
          maxLength: 20
          example: "TOKEN NOT VALID"
  responses:
    UnauthorizedError:
      description: Unauthorized - Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequestError:
      description: Bad request - requestBody not valid or empty
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    OperationNotAllowed:
      description: Forbidden - the operation is not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Not found - the content requested was not found in the database
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    BearerAuth:
      description:  |
        Define the security scheme for the api, in this case it's used Bearer Token,<br>
        and the token correspond to the usr_id
      type: http
      scheme: bearer
      bearerFormat: plain
paths:
...
  /users/{usr_id}:                                                #Endpoint di un utente
    parameters:
      - $ref: "#/components/parameters/usr_id"
    
    get:
      description:  |
        Retrive the photo and name of an user by its usr_id,
        useful for when an user want se info of other user in
        a chat or group
      summary: Get user info

      tags: ["User operation"]
      operationId: getUserInfo

      security:
        - BearerAuth: []
      
      responses:
        "200":
          description:  |
            Information retrived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              
              example:
                usrId: "ofnarc754"
                userName: "franco"
                userPhoto: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
        
        "400":
          $ref: "#/components/responses/BadRequestError"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "404":
          $ref: "#/components/responses/NotFoundError"

        "500":
          $ref: "#/components/responses/InternalError"
...
  /chats:                                                         #Endpoint delle chat
    get:
      description:  |
        Return an array that contain all the chats of a single user
      summary: List the chat of the user
      
      tags: ["Chat operation"]
      operationId: getMyConversations

      security:
        - BearerAuth: []

      responses:
        "200":
          description:  |
            The response contain a json object composed by an array with all(32 for now) the user chat
          content:
            application/json:
              schema:
                description: List with all user chat
                type: object
                properties:
                  chats:
                    description: user chats

                    type: array
                    items:
                      $ref: "#/components/schemas/Chat"
                    
                    minItems: 0
                    maxItems: 32
              
              example:
                chats:
                  - chat_id: 234
                    chatName: "Chat Uni"
                    chatPhoto: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                    isGroup: true
                    participants: ["ofnarc754", "arima53", "odfn34"]

                  - chat_id: 254
                    chatName: "Roberto"
                    chatPhoto: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                    isGroup: false
                    participants: ["ofnarc754", "ordigro293"]

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "404":
          $ref: "#/components/responses/NotFoundError"

        "500":
          $ref: "#/components/responses/InternalError"

    post:
      description:  |
        Consent the user to start a new chat with one or more user/s,
        the body need to have a message (the chat need to have at least 1 message)
      summary: Start new chat

      tags: ["Chat operation"]
      operationId: startNewChat

      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              description:  |
                The request body expect to retrieve the chat type, (name and photo if it's a group
                if isn't a group they can be empty) and it's participants minus the user that is starting the chat

              type: object
              properties:
                chatInfo:
                  description: |
                    Info of the chat
                  
                  type: object
                  properties:
                    chatName:
                      $ref: "#/components/schemas/Chat/properties/chatName"

                    chatPhoto:
                      $ref: "#/components/schemas/Chat/properties/chatPhoto"

                    isGroup:
                      $ref: "#/components/schemas/Chat/properties/isGroup"

                    participants:
                      $ref: "#/components/schemas/Chat/properties/participants"
                
                firstMessage:
                  description: |
                    First message required for creating the chat

                  type: object
                  
                  properties:
                    contentType:
                      $ref: "#/components/schemas/Message/properties/contentType"
                    content:
                      $ref: "#/components/schemas/Message/properties/content"

      responses:
        "201":
          description: Response with the chat info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

              example:
                chat_id: 344
                chatName: "Franco"
                chatPhoto: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                isGroup: false
                participants: ["ofnarc754", "arima53"]
        
        "400":
          $ref: "#/components/responses/BadRequestError"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "500":
          $ref: "#/components/responses/InternalError"
  /chats/{chat_id}:                                               #Endpoint di una chat
    parameters:
      - $ref: "#/components/parameters/chat_id"
    
    get:
      description:  |
        Retrive all the info of the chat
      summary: Get chat info

      tags: ["Chat operation"]
      operationId: getChatInfo

      security:
        - BearerAuth: []

      responses:
        "200":
          description: |
            Information retrived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

              example:
                chat_id: 344
                chatName: "Franco"
                chatPhoto: "iVBORw0KGgoAAAANSUhEUgAAAAUA..."
                isGroup: false
                participants: ["ofnarc754", "arima53"]

        "400":
          $ref: "#/components/responses/BadRequestError"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "403":
          $ref: "#/components/responses/OperationNotAllowed"

        "404":
          $ref: "#/components/responses/NotFoundError"

        "500":
          $ref: "#/components/responses/InternalError"

    put:
      description:  |
        Change group name
      summary: Change Group name

      tags: ["Group chat operation"]
      operationId: setGroupName

      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              description:  |
                The request body expects to receive the chatName field compatible
                with the one described in the chat schema 

              type: object
              properties:
                newGroupName:
                  $ref: "#/components/schemas/Chat/properties/chatName"

      responses:
        "202":
          description: Group chat name updated successfully
          content:
            application/json:
              schema:
                description:  |
                  Return the updated chat name

                type: object
                properties:
                  chatName:
                    $ref: "#/components/schemas/Chat/properties/chatName"

              example:
                chatName: "Chat belli"

        "400":
          $ref: "#/components/responses/BadRequestError"

        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "403":
          $ref: "#/components/responses/OperationNotAllowed"

        "404":
          $ref: "#/components/responses/NotFoundError"

        "500":
          $ref: "#/components/responses/InternalError"
...